Software backend: fix PaintedItem smooth xform
When a transform is applied to a QQuickPaintedItem, its 'smooth'
property should determine whether the transform is done with
antialising or not. This worked as expected with the default (rhi)
backend, but not with the software backend.

Since transforms do not cause item repaint, but are used when painting
the item's earlier painted and cached pixmap, the antialiasing is
determined not by QPainter's Antialiasing render hint, but by the
SmoothPixmapTransformation hint.

Fix by applying that hint according to the 'smooth' property value.

Fixes: QTBUG-124163
Change-Id: I7cc4e97ca768aa37f633e59d73796918f520f542
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit ad9b9a60de363e902ed51b8d4c0f90f1d432d3d0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>