ItemView: Avoid undesired repositioning in updateCurrent
When updateCurrent is called, it in turn calls updateHighlight.
updateHighlight already takes autoHighlight into consideration for some
of its logic, but still calls unconditionally into updateTrackedItem.
That in turn will trigger some movement of the list, to position the
tracked item, which is the highlight.

That movement is already only done for the case where the move reason is
setIndex. We add an additional condition to check that we really need to
track the highlight (either autoHighlight needs to be true, or
highlightRange is unequal to NoHighlightRange).

This does have the effect that the currentItem can be completely outside
of the viewport.

[ChangeLog][Important Behavior Change][QtQuick][ListView] If
highlightFollowsCurrentItem is set to false, and highlightRangeMode is
set to NoHighlightRange, then programatically setting the currentIndex
of the list will no longer scroll the view to that item.

Fixes: QTBUG-68527
Change-Id: I32aeec2621524d72981ec5d9b933d3f608d89390
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
(cherry picked from commit 60194389cd68f51376f611fe878a395c39e6f0eb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>