Android: fix handling of InputMethodQuery
Somewhere between Qt 6.5.1 and dev (backported to 6.5.4) something has
changed the behavior of QQuickItem::nextItemInFocusChain() or
QQuickItem::activeFocusOnTab(). I could not figure out the patch to
blame, because the actual code of these methods didn't change, but
as a result the loop to determine the next focus item in
QQuickTextInput::inputMethodQuery() was running forever.
Fix it by explicitly checking if the element changes between the loop
iterations, and if it does not, simply assing nullptr to it, acting as
if no focus item was found in the chain.

Fixes: QTBUG-123378
Pick-to: 6.7 6.5
Change-Id: I987f8ef2548cc116c2e69f5175d4af0bdb3ab3d0
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>