QQuickItem::setFocus: Don't return too early
a8be17a1472c6f504c0c40f68fdc13df035ac4b4 introduced a workaround for
setFocus returning too early when trying to set the active focus in
certain reparenting scenarios. In this workaround we check if, after
we try to setFocus() to the item and its ancestors, the item still
doesn't have activeFocus, then we try to set it again in the delivery
agent. However, it can happen that an item loses activeFocus at some
point during this execution. This is the case for example when the
item has an "onActiveFocusChanged" signal handler that calls
forceActiveFocus() on another item. So we can't always expect the
item to have activeFocus in the end.

As a fix, we should instead try to not return too early from
QQuickItem::setFocus. Currently, we return early if item->focus ==
focus. However, this is not enough of a check. When we set focus to
an item in Quick, we set the focus in a scope, an ancestor focus scope.
So we need to check both the focus flag AND that the item is the scoped
focus item of its nearest ancestor before we return early.

Amends a8be17a1472c6f504c0c40f68fdc13df035ac4b4.

Fixes: QTBUG-120670
Change-Id: I3f55f828d57fdc7d6f628813aa37930bdec0f994
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 23691ba87219ca82df3a75a74af4dab13d86a654)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit d49aa4022b48996f40a24932932896e80da228cd)
Reviewed-by: Doris Verria <doris.verria@qt.io>