richtext formatting of numbered and bullet lists (<ol> and <ul>)
Set the text color before drawing bullet points for numbered and bullet
point list, instead of using the color of the previous block.

It seems all other calls to addUnselectedGlyphs() are preceded by some
code to set the current color, so do it also for the bullet points.

Extend the manual test for syntaxhighlighter to add the reproducer.

Fixes: QTBUG-111337
Fixes: QTBUG-123855
Fixes: QTBUG-63741
Task-number: QTBUG-57833
Task-number: QTBUG-26612
Task-number: QTBUG-73859
Change-Id: I51a75152416e07c7f84f99b49e2a4b6ca72e78e9
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>