Mac style: draw dock widget titles and status bar without gradient
Fill with the "window" color of the palette as passed in through
the style option, which takes care of active vs inactive color
selection.

The title bar of a docked dockwidget is a bit lighter than the
main window title and unified toolbar background, but identical
to the color of the title of an undocked dock widget. This is now
the case for both dark mode and light modes.

Also remove the gradient fill from the status bar, just fill it
with the window brush.

Remove the now unused helper functions for the title bar
gradient.

This follows the equivalent commit in qtbase.

Task-number: QTBUG-92855
Change-Id: Ic887cc8c0d2f6b10f55029d54e74ce43110c03c0
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit 09a0ba13232420c87ae40e2fc4f0491d372b2164)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>