Mac style: draw MDI window title bar without gradient
Use a solid fill with the "button" color of the palette also for active
title bars, which is then the same color as the frame, and gives the
window a consistent look in both dark and light mode.

This matches the corresponding change in the widget style in
qtbase.

Task-number: QTBUG-123162
Change-Id: I95befbd0c51d641ceb033275a98c4e57a3564df5
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit b6bad50d0d33d873d27bea689e1b997b16924ec7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>