QQuickMenuBar: support changing the delegate
The application can change the delegate at run-time. The
problem is that a MenuBar can be of a mix of MenuBarItems
created from the delegate, and MenuBarItems added inline
in the QML code. And only the former should change if the
delegate changes.

To handle this, we need to know if an existing MenuBarItem
in the container was created from the old delegate before we
replace it with the new delegate. And to do this, we now
tag all the MenuBarItems we create with _qt_createdFromDelegate.

Moreover, the delegate can be null at startup, but change to
something "valid" later. And in the meantime, we need to
remember all the added menus. Since this information is
stored as MenuBarItems in the container, we need to temporarily
create hidden MenuBarItems when the delegate fails. This
ensures that the container alwyas contains all menus, regardless
of the delegate.

Change-Id: I29aee3c9967bfe0a77015ea902b6532fdcee387d
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>