Take into account inactive window when restoring focus after popup close
When popup is closing we try to restore the focus to the last
activeFocusItem of the window if the popup(item) has active focus.
However, it might happen that the window is inactive when the popup
closes, and the poup/popupItem don't have activeFocus. Take this into
account when deciding if popup should restore focus.
Since QQuickWindow doesn't store the focusObject when window loses
focus, check if popupItem is focusObject by going down the window's
subFocusItem hierarchy. Ideally the window wouldn't clear the focus
object when it loses focus but that's a bigger change that requires
further investigation.

Fixes: QTBUG-84976
Pick-to: 6.5
Change-Id: Ib896cb74fa85d3b6f894a80b84f3ffba0fb9c72e
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
(cherry picked from commit 2e520c6cae19f472b1053bd572bc488e8e048559)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>