QQuickMenu and MenuBar: add support for hiding and showing MenuBarItems
To hide a MenuBarItem from a native menu bar, we need to
remove and delete the native menu it maps to from QPA as
well (there is no way to hide it by other means). And when
we do so, we also need to remove and delete all the native
menu items inside the QQuickMenu as well. Otherwise the menu
items will continue to exist, but belong to a native menu
that has been removed and deleted.

Task-number: QTBUG-65461
Change-Id: I5767e4c08f64fc09246c61124acc856627a227ec
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>