DelegateChooser: React to changes of choice property (via QAIM)
Detect when a QAIM changes the value of the role affecting the choice
made by DelegateChooser and tell the view to recreate the delegate if
necessary.

For performance reasons, we don't want to track changes for non-QAIM
based models [1]; document this remaining limitation.

[1]: We don't have a convenient place where we could detect both the
"role" change and the affected index (which means that we would need to
invalidate everything, and possibly   way too often.). It might be
possible to refactor the code to allow for this, too, but that would
robably be a larger change, and not suitable for picking it back.

Fixes: QTBUG-75887
Change-Id: Ic91dc436e53b4280debf2698c76f2ead4bdf15d6
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit a407541a032de6f372206a0bc190775a0db950af)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>