qmldom: Add more file locations
On token is defined as colon token in the grammer file. If the flag
hasOnToken is set, then the colon token is actually on token.

Set file locations for Type *. This is needed for correctly highlighting
type names in the form of list<Item> etc.

Also add missing file locations for the if keyword, switch and for
keywords. These regions are required for semantic highlighting.

Task-number: QTBUG-120000
Change-Id: I8f6a8c0b317ff30422ccc8e1774ca3a80472a506
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>