QML: Introduce aliases for builtin types
This way we can express the relationship between qsizetype, qlonglong
and int, as well as the equivalents for long, ulong, char, qreal.

We consider aliases as equivalent to primitive types in
qmltyperegistrar. Therefore, you can only do this with types that don't
have a QMetaObject of their own. However, considering the use case, this
is not a big restriction.

When generating qmltypes, we propagate the aliases. When reading back
the qmltypes, we insert the resulting QQmlJSScope under all its separate
names. This makes all the aliases discernible, but actually equal types.

Since we now record the original type names in the qmltypes, this
changes a few test artifacts for qmltyperegistrar and qmllint.

Task-number: QTBUG-122582
Task-number: QTBUG-124294
Change-Id: I571684e3020b9bc30f8819bdc39389646fd1ff54
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>