qmltc: Test failure for unbound required property in inline component
`qmltc` should fail when it processes a QML file that instances an
inline component that has some top-level required properties without
setting them when creating the instance.

For example:

```
import QtQuick

Item {
   component IC: Item { required property int i }

   IC {
     // i needs to be set here as it cannot be set later
   }
}
```

To parse a QML file, `qmltc` uses `QQmlJSImportVisitor`, which will
produce an error on encountering such a case.

`qmltc` will in turn propagate the error and fail itself, thus acting
properly.

A new test was added to keep track of the behavior and ensure that
`qmltc` rejects a file when this erroneous form is encountered.

Task-number: QTBUG-120698
Change-Id: I1a1b7a3d55f3ba5dde7d04b014b7d6a9b53a546d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>