qmltc: Test failure for unbound required properties in component
`qmltc` should fail when it processes a QML file that defines a
component that has unbound inner-level required properties, similarly to
how it should generally fail when an unbound inner-level required
property is present, as those are not settable at a later phase.

To parse a QML file, `qmltc` uses `QQmlJSImportVisitor`, which will
produce an error on encountering such a case.

`qmltc` will in turn propagate the error and fail itself, thus acting
properly.

A new test was added to keep track of the behavior and ensure that
`qmltc` rejects a file when this erroneous form is encountered.

Task-number: QTBUG-120698
Change-Id: Iadb9648ea09951a1741679c9c27abe014876656b
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>