qmltc: Test fail for unbound outside required property in component
`qmltc` should fail when it processes a QML file that defines a
component that has unbound inner-level required properties imported from
an outside type.

For example:

```
Item {
    Component {
        id: mycomp

        Item {
            TypeWithRequiredProperty {}
        }
    }
}
```

Where `TypeWithRequiredProperty` defines at least one required property.

To parse a QML file, `qmltc` uses `QQmlJSImportVisitor`, which will
produce an error on encountering such a case.

`qmltc` will in turn propagate the error and fail itself, thus acting
properly.

A new test was added to keep track of the behavior and ensure that
`qmltc` rejects a file when this erroneous form is encountered.

Fixes: QTBUG-120698
Change-Id: I7f4dab3501c792b2b2a3d97b20ad416748a3d4a9
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>