qmltc: Move qlmtc_build_failures to tst_qmltc_qprocess
`qmltc_build_failures` uses the available CMake infrastructure to test
for expected build-level failures from a `qmltc` compilation.

This allows tracking failures over certain constructs that
`qmltc` should not support, but has an high amount of friction when
extended, as each "test case" needs to be its own whole project, and
doesn't provide an error-level granularity for the failures, so that the
success of a "test case" might depend on a failure that is different
from the intended one.

`tst_qmltc_qprocess` has the infrastructure required to run failing
`qmltc` builds in a leaner way, provides an easier extension story, based on
the general and well-known testlib-based structure and allows to
have tests with an error-level granularity.

Hence, remove `qmltc_build_failures` and move the existing "test case"
to `tst_qmltc_qprocess`.

Change-Id: Ib11227c0da1f6336e6537bea76011ccc0090039f
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>