QQDA: Make it easier to find the focus item when there's no active focus
QQuickWindow and QQuickDeliveryAgent only store the active focus item
and there was no trivial way of getting the item that should get focus
when window gains focus. This made it difficult to determine which is
the focus item of the window when the window is not active.
To make it easier, introduce targetFocusItem() that returns the item that
*should* get the active focus when the root focus scope gets active
focus, and use it in code that had to work-around this issue.

Task-number: QTBUG-121789
Change-Id: I4f054b256d0a64de4687d44509def1f478853d66
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>