QQuickText: consider DPR when drawing scalable inline images
Take into account the device pixel ratio when drawing scalable
inline images (such as SVG, SVGZ, PDF) in QML Text. Thus we can
improve the image rendering quality, especially for cases when the
image size is set explicitly.

Consider the following case:

Text {
    text: '<img src="foo.svg" width="16" height="16"/>'
}

In case the device pixel ratio is e.g. 1.5, we loaded the SVG file
into 16x16 QImage and after that we upscaled it as a raster to 24x24
with quality loss. Now we consider DPR when loading the image.

Note: this change improves only rendering for scalable images and
do not provide the support for "@2x" raster resources. But this can
be done in future changes.

Fixes: QTBUG-113040
Pick-to: 6.7
Change-Id: I8baada728e9acec5528553d0616ac37994ab80b6
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>