Android: Add QAndroidQuickView class that owns QAndroidViewSignalManager
Previously QAndroidViewSignalManager was a direct child
of the QQuickView. In JNI calls that we did not have a reference
to the signal manager, we had to use findChild to get access
to the instance of it.

We add QAndroidQuickView as a QQuickView subclass that owns
the QAndroidViewSignalManager instance.

Task-number: QTBUG-124488
Change-Id: I02ad10756a4af1ac82f130f349e7c8b6f563c70c
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>