QQmlObjectCreator: avoid extra detach on pendingBindings list
Do not detach on pendingBindings list when there is nothing to
remove in removePendingBinding() call. Previously it was always
detached due to the use of non-const iterators. To avoid this,
we can use QList::removeIf() which doesn't detach if there is
nothing to remove and also makes the code a bit clearer.

Change-Id: Ic2f8c30dcedcac5217e685c0dae55d7fb1615022
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>