QQuickState: Avoid potential extra QList detach
There might be potential extra QList detach due to the use of non-const
QList iterators in QQuickState::removeAllEntriesFromRevertList().
To avoid this, we can use QList::removeIf() which doesn't detach
if there is nothing to remove and also makes the code a bit clearer.

As a drive-by change also fix the code indentation.

Change-Id: I4a1761faaf6cf30fe1e93786ba3d094fd6e87ce9
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>