QQuickAbstractButton: setAction will share its Accessible object
In cases where a button acts as a view for an action, provide the
Action's accessibility information into the visual component so that it
can convey the action's information into the accessibility
infrastructure.
This is done by creating a Accessibility proxy object in the button
that will communicate the information from the action's accessibility
as part of the button. If the properties of the Accessibility object in
the button are modified, these values will prevail.

Fixes: QTBUG-123123
Change-Id: Ibeddcc918616d717bb1704177d482ff88bfe99ef
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>