More options for embedded images in QML generator
This adds a few options to the svgtoqml generator, to enable
more control over how embedded images are stored.

It adds the "keep-asset-paths" option. When this is set,
the generator will check if an image reference was originally
to a file on disk. Instead of duplicating this, the output
will contain a relative reference to the original file instead.

It adds the "asset-output-prefix" which overrides the default
prefix of output files.

And it adds the "asset-output-directory" which overrides the
default target directory. The default target directory has
also been changed to match the directory of the output QML
instead of always saving to the currently active directory.

In addition, this change makes the tool always output PNG
files, since JPEG is a lossy format and should never be preferred.
Previously, PNG was only preferred when the image had an alpha
channel.

Change-Id: I9a9b03eaea3ac511419789cfc5651d398ce42a5b
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>