Modernize Qt.labs.platform a bit
Separate its plugin from the base library and remove some generated
comments. We want to have the base library separate, so that its headers
become proper private headers and we can pass --private-headers to
qmltyperegistrar. Only with this option it will recognize the QPA
headers the module also uses.

Also, improve the readability of all the conditions and give it a name
that follows Qt's library naming convention.

Change-Id: Ieeaa4d69aafe4bb56f198deefbfa7c89e5a1f900
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>