qmlls: Add highlighting protocol with empty implementation
Add the non-functional semanticTokens protocol. Introduce the basic data
structures and the utility functions that is widely going to be used
during the rest of highlighting tasks.

When semanticTokens is requested, collectTokens visits the entire
domtree and accumulates the highlighting tokens. It is encoded to list
of integers as specified by the protocol at the end of collectTokens.

Add the utility tests in tst_qmlls_highlighting and module tests in
tst_qmlls_modules::semantichighlightingFull.

Task-number: QTBUG-1200000
Change-Id: I070251552a623eff4d4ff34e4201466b3ad2efab
Reviewed-by: Dmitrii Akshintsev <dmitrii.akshintsev@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>