Only apply QWindow::setVisibility() on first show or explicit request
As a result of 16023fc77c423a267fcc48894ff942e94cf35b86 we were no longer
calling QWindow::setVisible or setVisibility from the QQuickWindowQmlImpl
overrides, but instead deferred them until applyWindowVisibility().

As a consequence, if the QWindow::Visibility of the QQuickWindowQmlImpl
was left at AutomaticVisibility, we would always apply the window state
when making the window visible or hidden, which is a major regression.

We should only apply the window state on first show, if the window's
visibility is AutomaticVisibility, or on explicit requests to set the
window's visibility.

This means a window with visibility: Window.Maximized will initially
show as maximized, but if the user then resizes the window and then
hides and shows it again, the window will show in the normal window
state, instead of re-applying the Window.Maximized visibility, which
matches the pre-6.7.0 behavior.

Fixes: QTBUG-124363
Pick-to: 6.7.1
Change-Id: Idc2078c17bd56026425acbabdc40174fd12558ca
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 5da51e012094c6aad75238a04e0897d43a4975c6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>