semantichighlighting: add highlight for identifiers
Some identifiers cannot be handled in their containing dom type if they
contains dot in their names. Js variables that are not special types
are also highlighted here. Because an identifier can represent a few
different types, semantic analysis are performed to resolve their actual
types.

Also, adapt the binding handler's highlighting. A binding identifier can
be a property or a signal handler. In those cases, it should be
colorized as methods.

Pass built-in directory while creating the dom environment so that it
will behave the same as the code model.

Task-number: QTBUG-120000
Change-Id: Ifb8dda81ca7ef6f3150e68907c036387e32f279e
Reviewed-by: Dmitrii Akshintsev <dmitrii.akshintsev@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>