Add autotest for dialing a non-squared Dial
There was a bug where the Dial background would be aligned top-left on
some styles (Universal). This caused issues with mouse/touch input
because the Dial assumed that the Dial background was centered.

This was especially noticeable when the Dial was much wider than taller,
since the Dial background was unexpectedly at x == 0:
If you tried to move the handle to almost the middle (e.g. width/2 - 1)
it would move the handle towards its left side (but the mouse cursor was
on the right side of the visual Dial background because it was
left-aligned). Also, just clicking on the handle would cause it to jump
to a different position in most cases.

While investigating, this appears to have been fixed
(probably with bb7ba7667b4cf3565aa1849d08cc71b9ac011e77 ), so my fix is
not needed any longer. However, in order to avoid regressions, add the
autotest.

Change-Id: Iba9b599dd2945695a973e7de57f54f9907035a8d
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>