Add PlanarTransform object to facilitate 2D transforms
The existing function for creating a non-identity matrix4x4 requires
all of the 16 values to be specified. That is needlessly cumbersome in
the common case of a 2D (planar) transform. So add a new interface
singleton object to hold invokable methods that simplify specifying
such transforms in qml, avoiding exposing the full complexity of the
underlying 4x4 matrix in the api.

Change-Id: I8a94099da5c94f5483e6b46280c6101c6d3f6dc3
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>