qmlls: Add range support for semantic highlighting
Clients might ask only a range of the text document to be highlighted.
Then, we should perform highlighting for that particular range.
Introduce an optional data member m_range in the highlighting visitor.
If m_range has value, then highlight the current item only by checking
if it is overlapping with the requested range.

Task-number: QTBUG-120000
Change-Id: I635765d3d8831f7717edb96331db414f3813308f
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>