qmlls: Add delta support for semantic highlighting
If the document is too big, a client might not want to reperform all
highlights. Instead, they will request semanticTokens/delta to
understand what has changed on top of the latest data.

Server now holds a resultID string and the last encoded tokens. If the
client requests a delta with the same resultID then the difference is
computed and semanticeditstoken is populated. Else, servers' response will
contain the full semantic tokens.

Skip the module test for now as there is an unexpected behavior on
handling of variants in the language server api. This issue is tracked
with QTBUG-124870.

Task-number: QTBUG-120000
Change-Id: I5ea6d7a8f892d0be101dcc8f799937654618d79c
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>