macOS: Back off the render thread's updates when window is resized
Animators will normally do their updates exclusively on the render
thread, ensuring animations continue even if the main thread is
stalled. This is fine for normal operations, but during window
resizing we need to give the main thread a chance to present
drawables via displayLayer (which we plumb to exposeEvent),
so that the window size is continuously updated. If the render
thread is left to update and present as fast as it can, it will
starve the main thread from drawables, making resizing choppy.

We now detect if the window is being resized, and if so direct
the update to the main thread via requestUpdate, which gives
the QPA layer a chance to prioritize the resize updates over
the animations.

Pick-to: 6.8
Change-Id: Ice1e863a452d9eb4de72bae82db8dc91371ef329
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>