Android: Add Java QtAbstractItemModel as a QAIM wrapper
QtAbstractItemModel is a Java wrapper of QAbstractItemModel.
QtModelIndex is a Java wrapper of QModelIndex.

QtAbstractItemModel is an abstract Java class that wraps
commonly used public and protected functions of QAIM.

A QAndroidItemModelProxy instance can be created as a proxy
between the Java implementation and C++. The instance can be
used as a *QAbstractItemModel, for instance to be passed as a
QML property to the QQuickView instance, through QtQuickView.

The implementation also has a private QtAbstractItemModelProxy
Java class that exposes QAbstractItemModel* to the Java
QtAbstractItemModel instance.

QtModelIndex wraps a subset of QModelIndex APIs. As the ctor
of QModelIndex is private, we use row + column + internalId to
create new Java QtModelIndex instance when proxying from a
QtAbstractItemModel using the internal QtAndroidItemModelProxy.
When the proxying happens from *QAbstractItemModel to Java,
we use row + column + parent, as QAIM::index() should be used
instead of QAIM::createIndex().

This change also add required type conversions between C++
types and Java types.

[ChangeLog][Android] Added new public Java classes,
QtAbstractItemModel and QtModelIndex to wrap a subset of
their C++ counterpart functionalities, through internal proxies.

Task-number: QTBUG-124803
Change-Id: I2464d3d30592aeec8b678445c946e4d167dc4886
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>