Android: Add QtQmlComponent loading support to QtQuickView
A QQmlComponent component implementation, can deliver app library
name, module name, and the path to QML component to be loaded.

It also wraps set/get/connect methods of QtQuickView with protected
access. We can generate QQmlComponent concrete classes for selected
components at build time. Generated implementations, will have typed
set/get/connect methods, associated with properties and signals of the
QML component. Alternatively, it's possible for the user to subclass
QQmlComponent.

We add a new ctors to QtQuickView that takes only the Context. Later,
The user should instantiate and pass the QQmlComponent to
loadComponent(). QtQuickView gets required lib,module and component
from the implementation and loads the QML component by an internal
new QQuickView instance. If the there is already a QQuickView instance
available, it will be reused.

Task-number: QTBUG-124846
Change-Id: Ic29beb1b3faecd9f2e70621ddcc286a1a7659cb5
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>